==> project_2/01_count_data_preprocessing.qmd <==
---
title: "Project 2 Preprocessing"
author: "Jack Cunningham"
format: html
editor: visual
---

```{r, loading_packages}
library(readxl)
library(tidyverse)
```

Importing Count Data:

```{r}
count_data <- read_excel("Project3_S25_Data (1).xlsx", 
                         sheet = "ABS_LongCountData")
```

Getting treatments:

```{r}
count_data <- count_data |> 
  separate_wider_position(
    beaker,
    widths = c(temperature = 2, water = 1, replicate = 1),
    cols_remove = FALSE
  ) |> 
  pivot_wider(names_from = stage, values_from = count) |> 
  mutate(temperature = factor(temperature, c("20","28","33")),
         water = factor(water, c("R","A")))
  
```

Saving Count Data:

```{r}
save(count_data, file = "count_data.RData")
```
