==> project_2/03_count_data_modeling.qmd <==
---
title: "Count Data Modeling"
format: html
editor: visual
---

```{r}
library(lme4)
library(tidyverse)
load("count_data.RData")
```

# Trying to Model

```{r}
glm(data = count_data, polyp ~ water*time*temperature, family = poisson())
```

```{r}
library(rstatix)
```

```{r}
#res.aov <- anova_test(
#  data = count_data, dv = total, wid = beaker,
#  between = c(water,temperature),
#  within = c(time)
#)
#res.aov
```

```{r}
model_test <- lmer(bud ~ time*temperature*water + (1|beaker), data = count_data)
model_test
```

```{r}
model_test_poisson <- glmer(bud ~ (1|beaker) + (temperature|time) 
                            + water*temperature,
                           data = count_data, family = poisson())
summary(model_test_poisson)
```

```{r}
model_test_poisson_2 <- glmer(bud ~ (time|beaker) + water + temperature
                              + time:temperature + temperature:water,
                           data = count_data, family = poisson())
summary(model_test_poisson_2)
```

```{r}
library(Rmisc)
pred_count_data <- count_data
pred_count_data$bud <- predict(model_test_poisson_2)

pred_count_data.summary <- summarySE(pred_count_data, measurevar = "bud",
                                     groupvars = c("time","temperature","water")) |> 
  mutate(treatment = str_c(temperature, water, " "))
ggplot(pred_count_data.summary, aes(x = time, y = exp(bud), color = treatment)) +
  geom_errorbar(aes(ymin = exp(bud - se), ymax = exp(bud + se)), width = .5) +
  geom_line() +
  theme_bw() +
  geom_point(size = 3)
```

# Goal

Are trends across temperatures significant? Aka, do trends differ depending on temperature.

Are the differences within a temperature between microbe treatments significant? Aka, for a given temperature does the water treatment differ?

# Are Trends Across Temperatures Significant

```{r}
model_goal_one <- glmer(bud ~ (time|beaker) + temperature
                              + time:temperature,
                           data = count_data, family = poisson())
summary(model_goal_one)
```

```{r}
model_goal_one_2 <- model_goal_one <- glmer(bud ~ (1|beaker) + temperature
                              + time:temperature + water,
                           data = count_data, family = poisson())
summary(model_goal_one_2)
```

```{r}
pred_count_data <- count_data
pred_count_data$bud <- predict(model_goal_one_2)

pred_count_data.summary <- summarySE(pred_count_data, measurevar = "bud",
                                     groupvars = c("time","temperature","water")) |> 
  mutate(treatment = str_c(temperature, water, " "))
ggplot(pred_count_data.summary, aes(x = time, y = exp(bud), color = treatment)) +
  geom_errorbar(aes(ymin = exp(bud - se), ymax = exp(bud + se)), width = .5) +
  geom_line() +
  theme_bw() +
  geom_point(size = 3)
```

Yes, trends across temperatures are significant.

# Are Differences Within a Temperature Between Microbe Treatments Significant

```{r}
model_goal_two <- glmer(bud ~ (1+time|beaker) + temperature*water,
                           data = count_data, family = poisson())
summary(model_goal_two)
```

```{r}
pred_count_data <- count_data
pred_count_data$bud <- predict(model_goal_two)

pred_count_data.summary <- summarySE(pred_count_data, measurevar = "bud",
                                     groupvars = c("time","temperature","water")) |> 
  mutate(treatment = str_c(temperature, water, " "))
ggplot(pred_count_data.summary, aes(x = time, y = exp(bud), color = treatment)) +
  geom_errorbar(aes(ymin = exp(bud - se), ymax = exp(bud + se)), width = .5) +
  geom_line() +
  theme_bw() +
  geom_point(size = 3)
```

```{r}
model_goal_one_2 <- model_goal_one <- glmer(bud ~ (time|beaker) + temperature
                              + time:temperature + water,
                           data = count_data, family = poisson())
summary(model_goal_one_2)
```
