==> project_2/03_dna_data_confirmatory_analysis.qmd <==
---
title: "DNA Data Confirmatory Analysis"
author: "Jack Cunningham"
format: html
editor: visual
---

```{r}
load("DNA_data.RData")
```

# Are the differences between samples significantly different from one another

For the polyp DNA samples:

```{r}
polyp_Cq <- DNA_data |> 
  filter(Sample_Type == "Polyp")
```

```{r}
polyp_model <- lm(Cq ~ water*temperature + time, data = polyp_Cq)
summary(polyp_model)
anova(polyp_model)
```

Yes, they are different.

For the water DNA sample:

```{r}
water_Cq = DNA_data |> 
  filter(Sample_Type == "Water")
```

Model:

```{r}
water_model = lm(Cq ~ temperature*water, data = water_Cq)
anova(water_model)

water_model_means = emmeans(water_model, specs = ~ water*temperature) |>
  as.data.frame() |> 
  mutate(treatment = str_c(temperature, water, sep = " "))

water_model_means |> 
  ggplot(aes(x = treatment, y = emmean)) +
  geom_point() +
  geom_errorbar(aes(ymin = emmean - 2*SE, ymax = emmean + 2*SE, width = .1)) +
  labs(title = "Estimated Marginal Means Time 35")
```

Yes, they are significantly different.
